refactor: split pcg32 algorithm into its own module
Renames createPcg.ts to pcg32.ts so it mirrors the structure of
mulberry32.ts: each module exports `<variant>Advance(pcg, delta)`,
`<variant>Output(pcg)`, and a `create<Variant>` factory. The variant-
dispatched `nextState`, `getOutput`, `stepState`, `prevState`, and the
generic `randomInt` / `randomList` utilities now live in index.ts.
